Clean up pack objects manager usage
Because we always use a filter on the queryset, this view will never
raise DoesNotExist, so remove it
We never filter deleted packs on the view page, so the deleted templated
doesn't make sense
