+++ sounds/tests/test_sound.py
@@ -261,89 +261,6 @@ class ProfileNumSoundsTestCase(TestCase):
         self.assertEqual(pack.is_deleted, True)
 
 
-class PackNumSoundsTestCase(TestCase):
-
-    fixtures = ['licenses']
-
-    @mock.patch('sounds.models.delete_sounds_from_search_engine')
-    def test_create_and_delete_sounds(self, delete_sounds_from_search_engine):
-        N_SOUNDS = 5
-        user, packs, sounds = create_user_and_sounds(num_sounds=N_SOUNDS, num_packs=1)
-        pack = packs[0]
-        self.assertEqual(pack.num_sounds, 0)
-        for count, sound in enumerate(pack.sounds.all()):
-            sound.change_processing_state("OK")
-            sound.change_moderation_state("OK")
-            self.assertEqual(Pack.objects.get(id=pack.id).num_sounds, count + 1)  # Check pack has all sounds
-
-        sound_to_delete = sounds[0]
-        sound_to_delete_id = sound_to_delete.id
-        sound_to_delete.delete()
-        delete_sounds_from_search_engine.assert_called_once_with([sound_to_delete_id])
-        self.assertEqual(Pack.objects.get(id=pack.id).num_sounds, N_SOUNDS - 1)  # Check num_sounds on delete sound
-
-    def test_edit_sound(self):
-        N_SOUNDS = 1
-        user, packs, sounds = create_user_and_sounds(num_sounds=N_SOUNDS, num_packs=1)
-        pack = packs[0]
-        sound = sounds[0]
-        self.assertEqual(pack.num_sounds, 0)
-        sound.change_processing_state("OK")
-        sound.change_moderation_state("OK")
-        self.assertEqual(Pack.objects.get(id=pack.id).num_sounds, 1)  # Check pack has all sounds
-
-        self.client.force_login(user)
-        resp = self.client.post(reverse('sound-edit', args=[sound.user.username, sound.id]), {
-            '0-sound_id': sound.id,
-            '0-description': 'this is a description for the sound',
-            '0-name': sound.original_filename,
-            '0-tags': 'tag1 tag2 tag3',
-            '0-license': '3',
-            '0-new_pack': 'new pack name',
-            '0-pack': ''
-        })
-        self.assertRedirects(resp, reverse('sound', args=[sound.user.username, sound.id]))
-        self.assertEqual(Pack.objects.get(id=pack.id).num_sounds, 0)  # Sound changed from pack
-
-    def test_edit_pack(self):
-        user, packs, sounds = create_user_and_sounds(num_sounds=4, num_packs=2)
-        for sound in sounds:
-            sound.change_processing_state("OK")
-            sound.change_moderation_state("OK")
-        pack1 = packs[0]
-        pack2 = packs[1]
-        self.assertEqual(Pack.objects.get(id=pack1.id).num_sounds, 2)
-        self.assertEqual(Pack.objects.get(id=pack2.id).num_sounds, 2)
-
-        # Move one sound from one pack to the other
-        sound_ids_pack1 = [s.id for s in pack1.sounds.all()]
-        sound_ids_pack2 = [s.id for s in pack2.sounds.all()]
-        sound_ids_pack2.append(sound_ids_pack1.pop())
-        self.client.force_login(user)
-        resp = self.client.post(reverse('pack-edit', args=[pack2.user.username, pack2.id]), {
-            'pack_sounds': ','.join([str(sid) for sid in sound_ids_pack2]),
-            'name': 'Test pack 1 (edited)',
-            'description': 'A new description'
-        })
-        self.assertRedirects(resp, reverse('pack', args=[pack2.user.username, pack2.id]))
-        self.assertEqual(Pack.objects.get(id=pack1.id).num_sounds, 1)
-        self.assertEqual(Pack.objects.get(id=pack2.id).num_sounds, 3)
-
-        # Move one sound that had no pack
-        user, packs, sounds = create_user_and_sounds(num_sounds=1, num_packs=0, user=user, count_offset=5)
-        sound = sounds[0]
-        sound.change_processing_state("OK")
-        sound.change_moderation_state("OK")
-        resp = self.client.post(reverse('pack-edit', args=[pack2.user.username, pack2.id]), {
-            'pack_sounds':
-                ','.join([str(snd.id) for snd in Pack.objects.get(id=pack2.id).sounds.all()] + [str(sound.id)]),
-            'name': 'Test pack 1 (edited again)',
-            'description': 'A new description'
-        })
-        self.assertRedirects(resp, reverse('pack', args=[pack2.user.username, pack2.id]))
-        self.assertEqual(Pack.objects.get(id=pack1.id).num_sounds, 1)
-        self.assertEqual(Pack.objects.get(id=pack2.id).num_sounds, 4)
-        self.assertEqual(Sound.objects.get(id=sound.id).pack.id, pack2.id)
 
 
 class SoundViewsTestCase(TestCase):
