--- templates/sounds/pack_deleted.html
@@ -1,8 +1,0 @@
-{% extends "simple_page.html" %}
-
-{% block title %}Deleted pack{% endblock %}
-{% block page-title %}Deleted pack{% endblock %}
-
-{% block page-content %}
-    <p class="text-grey">This pack has been deleted.</p>
-{% endblock %}
